Rename hrit_msg reader to seviri_hrit
