+++ satpy/readers/__init__.py
@@ -51,6 +51,7 @@ OLD_READER_NAMES = {
     'hrit_electrol': 'electrol_hrit',
     'fci_fdhsi': 'fci_l1c_fdhsi',
     'hrit_goes': 'goes-imager_hrit',
+    'hrit_msg': 'seviri_hrit',
     'scmi_abi_l1b': 'abi_l1b_scmi',
 }
