+++ satpy/tests/reader_tests/test_ahi_hrit.py
