+++ satpy/tests/reader_tests/test_goes_imager_hrit.py
